+++ addon/appModules/xmplay/dialogs.py
@@ -9,6 +9,7 @@ import ui
 import wx
 
 from globalPlugins.xmplayAccessibility.configuration import get_setting
+from globalPlugins.xmplayAccessibility.localization import _
 
 from .backend import (
 	PLAYBACK_PAUSED,